StringifyValue moved to neptune.types
