+++ src/neptune/types/atoms/string.py
@@ -27,8 +27,8 @@ from neptune.internal.utils import (
     is_stringify_value,
     verify_type,
 )
-from neptune.internal.utils.stringify_value import StringifyValue
 from neptune.types.atoms.atom import Atom
+from neptune.types.stringify_value import StringifyValue
 
 if TYPE_CHECKING:
     from neptune.types.value_visitor import ValueVisitor
