+++ src/neptune/types/series/file_series.py
@@ -23,9 +23,9 @@ from typing import (
 
 from neptune.internal.utils import is_collection
 from neptune.internal.utils.logger import logger
-from neptune.internal.utils.stringify_value import extract_if_stringify_value
 from neptune.types import File
 from neptune.types.series.series import Series
+from neptune.types.stringify_value import extract_if_stringify_value
 
 if TYPE_CHECKING:
     from neptune.types.value_visitor import ValueVisitor
