+++ src/neptune/utils.py
@@ -23,7 +23,7 @@ from typing import (
     Union,
 )
 
-from neptune.internal.utils.stringify_value import StringifyValue
+from neptune.types.stringify_value import StringifyValue
 
 
 def stringify_unsupported(value: Any) -> Union[StringifyValue, Mapping, List, Tuple]:
